Table2 renamed to Table.
